feat: first implementation supporting nickel configurations
