--- api/v0/package/package.proto
@@ -1,46 +1,0 @@
-syntax = "proto3";
-
-package vorpal.package.v0;
-
-service PackageService {
-    rpc Prepare (stream PackagePrepareRequest) returns (stream PackagePrepareResponse);
-    rpc Build (PackageBuildRequest) returns (stream PackageBuildResponse);
-}
-
-message PackagePrepareRequest {
-    bytes source_data = 1;
-    string source_hash = 2;
-    string source_name = 3;
-    string source_signature = 4;
-}
-
-message PackagePrepareResponse {
-    string log_output = 1;
-}
-
-message PrepareBuildPackage {
-    string hash = 1;
-    string name = 2;
-}
-
-enum PackageBuildSystem {
-    UNKNOWN_SYSTEM = 0;
-    AARCH64_LINUX = 1;
-    AARCH64_MACOS = 2;
-    X86_64_LINUX = 3;
-    X86_64_MACOS = 4;
-}
-
-message PackageBuildRequest {
-    PackageBuildSystem build_system = 1;
-    map<string, string> build_environment = 2;
-    optional string build_image = 3;
-    repeated PrepareBuildPackage build_packages = 4;
-    string build_script = 5;
-    string source_hash = 6;
-    string source_name = 7;
-}
-
-message PackageBuildResponse {
-    string log_output = 1;
-}
