--- api/v0/store/store.proto
@@ -1,28 +1,0 @@
-syntax = "proto3";
-
-package vorpal.store.v0;
-
-service StoreService {
-    rpc Fetch (StorePath) returns (stream StoreFetchResponse);
-    rpc Path (StorePath) returns (StorePathResponse);
-}
-
-enum StorePathKind {
-    UNKNOWN = 0;
-    SOURCE = 1;
-    PACKAGE = 2;
-}
-
-message StorePath {
-    StorePathKind kind = 1;
-    string hash = 2;
-    string name = 3;
-}
-
-message StoreFetchResponse {
-    bytes data = 1;
-}
-
-message StorePathResponse {
-    string uri = 1;
-}
