--- src/lib.rs
@@ -1,8 +1,0 @@
-pub mod api {
-    tonic::include_proto!("vorpal.config.v0");
-    tonic::include_proto!("vorpal.package.v0");
-    tonic::include_proto!("vorpal.store.v0");
-}
-pub mod notary;
-pub mod service;
-pub mod store;
