--- src/store/temps.rs
@@ -1,17 +1,0 @@
-use crate::store::paths;
-use anyhow::Result;
-use std::path::PathBuf;
-use tokio::fs;
-use tokio::fs::File;
-
-pub async fn create_dir() -> Result<PathBuf> {
-    let temp_dir_path = paths::get_temp_path();
-    fs::create_dir(&temp_dir_path).await?;
-    Ok(temp_dir_path.canonicalize()?)
-}
-
-pub async fn create_file(extension: &str) -> Result<PathBuf> {
-    let temp_file_path = paths::get_temp_path().with_extension(extension);
-    File::create(&temp_file_path).await?;
-    Ok(temp_file_path.canonicalize()?)
-}
