+++ store/src/hashes.rs
