+++ store/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod archives;
+pub mod hashes;
+pub mod paths;
+pub mod temps;
