+++ worker/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "vorpal-worker"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "vorpal-worker"
+path = "src/main.rs"
+
+[dependencies]
+# async_zip = { version = "0", features = ["deflate", "tokio"] }
+# clap = { version = "4", features = ["derive"] }
+# futures-lite = { version = "2", default-features = false, features = ["std"] }
+# nickel-lang-core = "0"
+# petgraph = "0"
+# rand = "0"
+# serde_json = "1"
+# sha256 = "1"
+# tokio-util = { version = "0", features = ["compat"] }
+# tracing-subscriber = "0"
+# walkdir = "2"
+anyhow = "1"
+async-compression = { version = "0", features = ["all"] }
+bollard = { version = "0", features = ["buildkit"] }
+clap = { version = "4", features = ["derive"] }
+dirs = "5"
+git2 = "0"
+hex = "0"
+infer = "0"
+prost = "0"
+reqwest = { version = "0", features = ["json"] }
+rsa = { version = "0", features = ["sha2"] }
+tokio = { version = "1", features = ["full"] }
+tokio-stream = "0"
+tokio-tar = "0"
+tonic = "0"
+tracing = "0"
+tracing-subscriber = "0"
+url = "2"
+uuid = { version = "1", features = ["v7"] }
+vorpal-notary = { path = "../notary" }
+vorpal-schema = { path = "../schema" }
+vorpal-store = { path = "../store" }
+
+[build-dependencies]
+tonic-build = "0"
