+++ worker/api/v0/store/store.proto
@@ -1,0 +1,29 @@
+syntax = "proto3";
+
+package vorpal.store.v0;
+
+service StoreService {
+    rpc Exists (StoreRequest) returns (StoreExistsResponse);
+    rpc Pull (StoreRequest) returns (stream StorePullResponse);
+}
+
+enum StoreKind {
+    UNKNOWN = 0;
+    INPUT = 1;
+    PACKAGE = 2;
+    SOURCE = 3;
+}
+
+message StoreRequest {
+    StoreKind kind = 1;
+    string hash = 2;
+    string name = 3;
+}
+
+message StoreExistsResponse {
+    bool exists = 1;
+}
+
+message StorePullResponse {
+    bytes data = 1;
+}
